Added pool of paths

+++ include/my_types.hh
@@ -6,9 +6,9 @@
 #include <boost/integer_traits.hpp>
 
 typedef int16_t   node_t;
-typedef int32_t   edge_t;
+typedef int16_t   edge_t;
 typedef double    cost_t;
-typedef double    resource_t;
+typedef int16_t   resource_t;
 typedef double    dist_t;
 
 #include <vector>
